requirements.txt
pillow
pytesseract
pdfplumber
langchain
langgraph
langdetect